texk/dvipdfm-x: General cleanup: No reason to use long for 32-bit integers
git-svn-id: svn://tug.org/texlive/trunk/Build/source@35176 c570f23f-e606-0410-a88d-b1316a301751

+++ texk/dvipdfm-x/dvipdfmx.c
@@ -747,7 +747,7 @@ error_cleanup (void)
 static void
 do_dvi_pages (void)
 {
-  long     page_no, page_count, i, step;
+  int      page_no, page_count, i, step;
   double   page_width, page_height;
   double   init_paper_width, init_paper_height;
   pdf_rect mediabox;
